+++ WcfBankingService/operation/operations/Withdraw.cs
@@ -5,7 +5,7 @@ namespace WcfBankingService.operation.operations
 {
     public class Withdraw : BankOperation, IBankCommand
     {
-        public Withdraw(string operationTitle, BigInteger amount, string source) : base(operationTitle, amount, source)
+        public Withdraw(string operationTitle, decimal amount, string source) : base(operationTitle, amount, source)
         {
         }
